# Another performance optimization for Zen4 + refactoring