@ikawrakow 2024-05-23
> > I noticed that my AVX2 implemetation of Q8_K quantization
> 
> Could you go into more detail on this? Did I accidentally remove that in the last sync? The last sync was 30k lines (a lot of upstream development for 2 weeks!) so sometimes things get lost in translation. The `[jart]` comment markers help me avoid doing that by mistake.

I had added `AVX2` implementation for quantizing to `Q8_K` in the initial PR, see `quantize_row_q8_K` in https://github.com/Mozilla-Ocho/llamafile/pull/394/files. I did it that way because I didn't want to fool around with Georgi's single-threaded `  GGML_TASK_TYPE_INIT`. But I actually like what you have done better. Once `GGML_TASK_TYPE_INIT` is multi-threaded, there is no performance benefit from vectorizing the quantization to `Q8_K` (I measured with and without `Q8_K` AVX2 implementation and it made no measurable difference on my computer).